File strucure reorganization, fit optimization
