+++ acma/Scripts/numba
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# -*- coding: UTF-8 -*-
+from __future__ import print_function, division, absolute_import
+
+from numba.misc.numba_entry import main
+
+if __name__ == "__main__":
+    main()
