+++ acma/Scripts/pycc
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+from numba.pycc import main
+main()
