+++ data/AP424/chiT1000.csv
@@ -1,0 +1,14 @@
+[Data]
+
+Comment,Time Stamp (sec),Temperature (K),Magnetic Field (Oe),Moment (emu),M. Std. Err. (emu),Transport Action,Averaging Time (sec),Frequency (Hz),Peak Amplitude (mm),Center Position (mm),Lockin Signal' (V),Lockin Signal (V),Range,M. Quad. Signal (emu),AC Moment (emu),AC M. Std Err. (emu),AC Phase (deg),AC Phase Std. Err. (deg),AC Susceptibility (emu/Oe),AC Suscept. Std Err. (emu/Oe),AC X' (emu/Oe),AC X' Std Err. (emu/Oe),AC X'' (emu/Oe),AC X'' Std Err. (emu/Oe),AC Drive (Oe),AC Frequency (Hz),AC Averaging Time (sec),AC Cycles,AC Range,AC Measure Type,AC Signal' (V),AC Signal'' (V),AC Trim Coil Ratio,AC Trim Coil Phase,Min. Temperature (K),Max. Temperature (K),Min. Field (Oe),Max. Field (Oe),Mass (grams),Motor Lag (deg),Pressure (Torr),Measure Count,Measurement Number,SQUID Status (code),Motor Status (code),Measure Status (code),Motor Current (amps),Motor Temp. (C),Temp. Status (code),Field Status (code),Chamber Status (code),Chamber Temp (K),Redirection State,Evercool Status,Average Temp (K),Rotation Angle (deg),Rotator state,DC Moment Fixed Ctr (emu),DC Moment Err Fixed Ctr (emu),DC Moment Free Ctr (emu),DC Moment Err Free Ctr (emu),DC Fixed Fit,DC Free Fit,DC Calculated Center (mm),DC Calculated Center Err (mm),DC Scan Length (mm),DC Scan Time (s),DC Number of Points,DC Squid Drift,DC Min V (V),DC Max V (V),DC Scans per Measure,Map 01,Map 02,Map 03,Map 04,Map 05,Map 06,Map 07,Map 08,Map 09,Map 10,Map 11,Map 12,Map 13,Map 14,Map 15,Map 16
+
+,3826817369.37128,1.80166286230087,999.746765136719,,,6,,,,35.210693359375,,,100,,,,,,,,,,,,,,,,,,,,,,1.79721021652222,1.80225837230682,999.746765136719,999.746765136719,,,2.4124653339386,-99999999,1,0,0,0,0.0145721435546875,33.2355842590332,1,1,1,1.80136013031006,0,1816139040,1.80046590169271,,,0.00328224963844721,0.000161946688354772,0.00328322667024605,0.000161948744799096,0.94536966085434,0.950845599174499,35.1189956665039,0.186689734458923,30,3,200,-0.000653853989206254,-0.139781266450882,0.0208577774465084,1,,,,,,,,,,,,,,,,
+
+,3826817384.1719,1.94431298971176,999.746765136719,,,6,,,,35.210693359375,,,100,,,,,,,,,,,,,,,,,,,,,,1.83009827136993,2.03966355323791,999.746765136719,999.746765136719,,,2.59143090248108,-99999999,1,0,0,0,0.0138092041015625,33.2355842590332,6,1,1,2.05852770805359,0,1816138912,1.93234705924988,,,0.00298744216563741,0.000147482443162713,0.00298838107424731,0.000147484394249391,0.936062633991241,0.940938532352447,35.1164703369141,0.186757311224937,30,3,200,-0.000647992012090981,-0.129380598664284,0.0131525453180075,1,,,,,,,,,,,,,,,,
+
+,3826817393.37729,2.18272852897644,999.746765136719,,,6,,,,35.210693359375,,,100,,,,,,,,,,,,,,,,,,,,,,2.06345891952515,2.28655862808228,999.746765136719,999.746765136719,,,2.74497604370117,-99999999,1,0,0,0,0.0138092041015625,32.7468223571777,6,1,1,2.30199813842773,0,1816139072,2.17020069228278,,,0.00265303802891444,0.00013095368359317,0.00265381101653638,0.00013095515568577,0.938500761985779,0.943220019340515,35.1197776794434,0.186772465705872,30,3,200,-0.000486005301354453,-0.110050767660141,0.0158248506486416,1,,,,,,,,,,,,,,,,
+
+,3826817402.6465,2.42910003662109,999.746765136719,,,6,,,,35.210693359375,,,100,,,,,,,,,,,,,,,,,,,,,,2.31165099143982,2.52265191078186,999.746765136719,999.746765136719,,,2.88886761665344,-99999999,1,0,0,0,0.0138092041015625,33.2355842590332,6,1,1,2.54654908180237,0,1816139072,2.41657441854477,,,0.00239311468386424,0.000118116006619385,0.00239377160528506,0.000118117174164727,0.939592480659485,0.944127857685089,35.1223106384277,0.186786025762558,30,3,200,-0.000377680436940864,-0.103354126214981,0.00976906530559063,1,,,,,,,,,,,,,,,,
+
+,3826817416.3577,2.79926037788391,999.746765136719,,,6,,,,35.210693359375,,,10,,,,,,,,,,,,,,,,,,,,,,2.55103635787964,3.03157305717468,999.746765136719,999.746765136719,,,3.15859651565552,-99999999,1,0,0,0,0.0138092041015625,33.7243423461914,6,1,1,3.04748439788818,0,1816138976,2.78775314723744,,,0.00198316281803375,9.78696988951983E-5,0.0019836344711494,9.78705109574226E-5,0.941729605197906,0.945797741413116,35.1283874511719,0.186733409762382,30,3,200,-0.00235981936566532,-0.842540800571442,0.140284076333046,1,,,,,,,,,,,,,,,,
+
