+++ main.py
@@ -10,8 +10,6 @@ if __name__ == "__main__":
     MainWindow = QtWidgets.QMainWindow()
     ui = MainPage(MainWindow)
 
-    
-
     MainWindow.showMaximized()
 
     sys.exit(app.exec_())
