+++ requirement.txt
@@ -4,3 +4,4 @@ pandas
 numpy 
 pandas
 scipy
+numba
